chore(daemon): split task impls
